Revert "fixed some visual studio warnings"
This reverts commit 3c103e7e2fedffc6953d82a8c1c2f78a87f41309.

+++ Src/Graphics/ModelImageGenerator.cpp
@@ -59,7 +59,7 @@ PGE::Texture* ModelImageGenerator::generate(const PGE::String& model, float scal
         // Further optimized if in initialization step, requires Model shader to be loaded at that point.
         gfxRes->setCameraUniforms(cam);
 
-        position = position * (float)texSize;
+        position = position * texSize;
         mi->setPosition(PGE::Vector3f(position.x, position.y, 0.f));
         mi->setRotation(rotation);
         mi->setScale(PGE::Vector3f(scale * texSize));
